requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
pytest
